+++ src/Credfeto.DotNet.Repo.Tools.TemplateUpdate.Tests/DependencyInjectionTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Net.Http;
+using Credfeto.ChangeLog;
 using Credfeto.DotNet.Repo.Tools.Build.Interfaces;
 using Credfeto.DotNet.Repo.Tools.DotNet.Interfaces;
 using Credfeto.DotNet.Repo.Tools.Git.Interfaces;
@@ -70,6 +71,9 @@ public sealed class DependencyInjectionTests : DependencyInjectionTestsBase
             .AddMockedService<IReleaseGeneration>()
             .AddMockedService<ITrackingCache>()
             .AddMockedService<IDotNetFilesDetector>()
+            .AddMockedService<IChangeLogDetector>()
+            .AddMockedService<IChangeLogUpdater>()
+            .AddMockedService<IChangeLogLanguageFactory>()
             .AddTemplateUpdate();
     }
 }